로스트시티 평가 로직 분리
